Restructure random_generator
