+++ kornia/augmentation/random_generator/__init__.py
@@ -1,2 +1,2 @@
-from .random_generator import *
-from .random_generator3d import *
+from ._2d import *
+from ._3d import *
